fix(parity): align iterative BLINK and FarmCPU workflows
- reproduce GAPIT 3.5 prefix-BIC, previous-QTN retention, and SUB reward behavior
- align cofactor GLM and mixed-model edge-case statistics
- add cross-language workflow coverage for BLINK, FarmCPU, MLM/EMMAX, and gBLUP
- align NumPy scalar annotations with runtime types without runtime coercion

+++ pygapit/gwas/mlm.py
@@ -99,7 +99,7 @@ def compress_kinship(
     if n_groups <= 1:
         # All in one group → GLM equivalent
         Z = np.ones((n, 1)) / n
-        K_c = np.array([[float(K.mean())]])
+        K_c = np.array([[K.mean()]])
         return K_c, Z
 
     try:
